requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cycler==0.10.0
daphne==2.5.0
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kiwisolver==1.2.0
matplotlib==3.3.2
misaka==2.1.1
msgpack==1.0.0
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
razorpay==1.2.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
twilio==6.45.4
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.2
